requirements.txt
numpy
google-cloud-aiplatform
langchain-google-vertexai
langchain
langchain-google-genai 
langchain-google-community
unstructured[pdf]
streamlit